test with double registration. NOTE: not yet possible to define both double and float due to functor typename

+++ src/registration.cpp
@@ -109,6 +109,6 @@ namespace duna
     }
 
     // Instantiations
-    template class Registration<pcl::PointXYZ, pcl::PointXYZ, float>;
-    // template class Registration<pcl::PointXYZ, pcl::PointXYZ, double>;
+    // template class Registration<pcl::PointXYZ, pcl::PointXYZ, float>;
+    template class Registration<pcl::PointXYZ, pcl::PointXYZ, double>;
 }
